WIP: move LogViewer class into util module and add tests
